+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## 7.13.0
+
 ### Added
 - Support `on-demand-install-capable` application https://github.com/tuist/XcodeProj/pull/554 by @d-date
 - Add RemotePath to RemoteRunnable https://github.com/tuist/XcodeProj/pull/555 by @kwridan